+++ src/FilamentUserAttributes.php
@@ -2,6 +2,8 @@
 
 namespace Luttje\FilamentUserAttributes;
 
+use Filament\Forms\Components\Component;
+use Filament\Forms\Form;
 use Illuminate\Support\Facades\File;
 use Luttje\FilamentUserAttributes\Contracts\HasUserAttributesConfigContract;
 use Luttje\FilamentUserAttributes\Contracts\HasUserAttributesContract;
@@ -105,5 +107,115 @@ class FilamentUserAttributes
             ->toArray();
 
         return $resources;
+    }
+
+    /**
+     * Helper function to get label for a component.
+     */
+    private static function getComponentLabel($component, ?string $parentLabel = null): string
+    {
+        $label = $component->getLabel();
+
+        if (!empty($label)) {
+            return $parentLabel ? ($parentLabel . ' > ' . $label) : $label;
+        }
+
+        return $parentLabel ?? '';
+    }
+
+    /**
+     * Gets all components and child components as a flat array of names with labels
+     */
+    public static function getAllFieldComponents(array $components, ?string $parentLabel = null): array
+    {
+        $namesWithLabels = [];
+
+        foreach ($components as $component) {
+            $label = static::getComponentLabel($component, $parentLabel);
+
+            if ($component instanceof \Filament\Forms\Components\Field) {
+                $namesWithLabels[] = [
+                    'name' => $component->getName(),
+                    'label' => $label,
+                ];
+            }
+
+            if ($component instanceof Component) {
+                $namesWithLabels = array_merge(
+                    $namesWithLabels,
+                    static::getAllFieldComponents(
+                        $component->getChildComponents(),
+                        $label
+                    )
+                );
+            }
+        }
+
+        return $namesWithLabels;
+    }
+
+    /**
+     * Search the components and child components until the component with the given name is found,
+     * then add the given component after it.
+     */
+    public static function addComponentAfterComponent(array $components, string $siblingComponentName, bool $before, Component $componentToAdd, ?string $parentLabel = null): array
+    {
+        $newComponents = [];
+
+        foreach ($components as $component) {
+            $label = static::getComponentLabel($component, $parentLabel);
+
+            $newComponents[] = $component;
+
+            if ($component instanceof \Filament\Forms\Components\Field
+            && $label === $siblingComponentName) {
+                if (!$before) {
+                    $newComponents[] = $componentToAdd;
+                } else {
+                    array_splice($newComponents, count($newComponents) - 1, 0, [$componentToAdd]);
+                }
+            }
+
+            if ($component instanceof Component) {
+                $childComponents = static::addComponentAfterComponent(
+                    $component->getChildComponents(),
+                    $siblingComponentName,
+                    $before,
+                    $componentToAdd,
+                    $label
+                );
+
+                $component->childComponents($childComponents);
+            }
+        }
+
+        return $newComponents;
+    }
+
+    /**
+     * Merges the custom fields into the form.
+     */
+    public static function mergeCustomFormFields(Form $form, array $components, string $resource): void
+    {
+        $customFields = FilamentUserAttributes::getUserAttributeComponents($resource);
+
+        $appendComponents = [];
+
+        foreach ($customFields as $customField) {
+            if ($customField['ordering']['sibling'] === null) {
+                $appendComponents[] = $customField['component'];
+                continue;
+            }
+            $components = self::addComponentAfterComponent(
+                $components,
+                $customField['ordering']['sibling'],
+                $customField['ordering']['before'],
+                $customField['component']
+            );
+        }
+
+        $components = array_merge($components, $appendComponents);
+
+        $form->components($components);
     }
 }
