+++ src/Traits/HasUserAttributesResource.php
@@ -44,26 +44,8 @@ trait HasUserAttributesResource
 
         $components = self::resourceForm($form)
             ->getComponents();
-        $customFields = FilamentUserAttributes::getUserAttributeComponents(self::class);
 
-        $appendComponents = [];
-
-        foreach ($customFields as $customField) {
-            if ($customField['ordering']['sibling'] === null) {
-                $appendComponents[] = $customField['component'];
-                continue;
-            }
-            $components = self::addComponentAfterComponent(
-                $components,
-                $customField['ordering']['sibling'],
-                $customField['ordering']['before'],
-                $customField['component']
-            );
-        }
-
-        $components = array_merge($components, $appendComponents);
-
-        $form->components($components);
+        FilamentUserAttributes::mergeCustomFormFields($form, $components, self::class);
 
         return $form;
     }
@@ -81,7 +63,7 @@ trait HasUserAttributesResource
         $components = self::resourceForm($form)
             ->getComponents();
 
-        return static::getAllFieldComponents($components);
+        return FilamentUserAttributes::getAllFieldComponents($components);
     }
 
     /**
@@ -90,126 +72,6 @@ trait HasUserAttributesResource
     public static function getColumnsForOrdering(): array
     {
         return [];
-    }
-
-    // /**
-    //  * Gets all components and child components as a flat array of names with labels
-    //  */
-    // public static function getAllFieldComponents(array $components, ?string $parentLabel = null): array
-    // {
-    //     $namesWithLabels = [];
-
-    //     foreach ($components as $component) {
-    //         $label = $component->getLabel();
-
-    //         if (!empty($label)) {
-    //             $label = $parentLabel ? ($parentLabel . ' > ' . $component->getLabel()) : $component->getLabel();
-    //         } else {
-    //             $label = $parentLabel;
-    //         }
-
-    //         if ($component instanceof \Filament\Forms\Components\Field) {
-    //             $namesWithLabels[] = [
-    //                 'name' => $component->getName(),
-    //                 'label' => $label,
-    //             ];
-    //         }
-
-    //         if ($component instanceof Component) {
-    //             $namesWithLabels = array_merge(
-    //                 $namesWithLabels,
-    //                 static::getAllFieldComponents(
-    //                     $component->getChildComponents(),
-    //                     $label
-    //                 )
-    //             );
-    //         }
-    //     }
-
-    //     return $namesWithLabels;
-    // }
-
-    /**
-     * Helper function to get label for a component.
-     */
-    private static function getComponentLabel($component, ?string $parentLabel = null): string
-    {
-        $label = $component->getLabel();
-
-        if (!empty($label)) {
-            return $parentLabel ? ($parentLabel . ' > ' . $label) : $label;
-        }
-
-        return $parentLabel ?? '';
-    }
-
-    /**
-     * Gets all components and child components as a flat array of names with labels
-     */
-    public static function getAllFieldComponents(array $components, ?string $parentLabel = null): array
-    {
-        $namesWithLabels = [];
-
-        foreach ($components as $component) {
-            $label = static::getComponentLabel($component, $parentLabel);
-
-            if ($component instanceof \Filament\Forms\Components\Field) {
-                $namesWithLabels[] = [
-                    'name' => $component->getName(),
-                    'label' => $label,
-                ];
-            }
-
-            if ($component instanceof Component) {
-                $namesWithLabels = array_merge(
-                    $namesWithLabels,
-                    static::getAllFieldComponents(
-                        $component->getChildComponents(),
-                        $label
-                    )
-                );
-            }
-        }
-
-        return $namesWithLabels;
-    }
-
-    /**
-     * Search the components and child components until the component with the given name is found,
-     * then add the given component after it.
-     */
-    public static function addComponentAfterComponent(array $components, string $siblingComponentName, bool $before, Component $componentToAdd, ?string $parentLabel = null): array
-    {
-        $newComponents = [];
-
-        foreach ($components as $component) {
-            $label = static::getComponentLabel($component, $parentLabel);
-
-            $newComponents[] = $component;
-
-            if ($component instanceof \Filament\Forms\Components\Field
-            && $label === $siblingComponentName) {
-                if (!$before) {
-                    $newComponents[] = $componentToAdd;
-                } else {
-                    array_splice($newComponents, count($newComponents) - 1, 0, [$componentToAdd]);
-                }
-            }
-
-            if ($component instanceof Component) {
-                $childComponents = static::addComponentAfterComponent(
-                    $component->getChildComponents(),
-                    $siblingComponentName,
-                    $before,
-                    $componentToAdd,
-                    $label
-                );
-
-                $component->childComponents($childComponents);
-            }
-        }
-
-        return $newComponents;
     }
 }
