+++ tests/Livewire/ConfiguredManageComponentTest.php
@@ -14,7 +14,7 @@ it('can render a form with configured user attributes', function () {
     // Pretend that the user configured what attributes exist for products
     $user->userAttributesConfigs()->create(
         [
-            'resource_type' => Product::class,
+            'resource_type' => ConfiguredManageComponent::class,
             // TODO: Use a scope for always set this:
             'owner_type' => get_class($user),
             'config' => [
